cambiado el dato de Integer a BigDecimal

+++ src/main/java/componentesMatematicos/AtributoCalculo.java
@@ -1,5 +1,7 @@
 package componentesMatematicos;
 
+import java.math.BigDecimal;
+
 import model.BaseDeDatos;
 import model.Empresa;
 
@@ -12,7 +14,7 @@ public class AtributoCalculo extends FactorLiteral{
 		return this.nombre;
 	}
 	
-	public Integer getValor(Integer unPeriodo, Empresa unaEmpresa, BaseDeDatos unaBaseDeDatos) {
+	public BigDecimal getValor(Integer unPeriodo, Empresa unaEmpresa, BaseDeDatos unaBaseDeDatos) {
 		return unaBaseDeDatos.valorDe(this.nombre, unPeriodo, unaEmpresa);
 		//return unaBaseDeDatos.buscarAtributo(this.nombre, unaEmpresa).getValor(unPeriodo, unaEmpresa, unaBaseDeDatos);
 	}
