+++ src/main/java/componentesMatematicos/Expresion.java
@@ -1,9 +1,11 @@
 package componentesMatematicos;
 
+import java.math.BigDecimal;
+
 import model.BaseDeDatos;
 import model.Empresa;
 
 public interface Expresion {
 	
-	public Integer getValor(Integer unPeriodo, Empresa unaEmpresa, BaseDeDatos unaBaseDeDatos);
+	public BigDecimal getValor(Integer unPeriodo, Empresa unaEmpresa, BaseDeDatos unaBaseDeDatos);
 }
